implement status for weather data retrieval and error handling

+++ lib/models/temperature.dart
@@ -8,9 +8,9 @@ class TemperatureModel extends Equatable {
   TemperatureModel({this.day, this.min, this.max});
 
   TemperatureModel.fromJson(Map<String, dynamic> json) {
-    day = json['day'];
-    min = json['min'];
-    max = json['max'];
+    day = json['day'].toDouble();
+    min = json['min'].toDouble();
+    max = json['max'].toDouble();
   }
 
   @override
